+++ NEWS.md
@@ -2,6 +2,8 @@
 ---------------------
 * Modified `abstract` function
   * New `dt` argument indexes significant results in `@pairs`
+* New `adjacent` function
+  * Uses `@pairs` slot to build an adjacency matrix
 
 ## propr 2.0.8
 ---------------------